refactor(core): single-home the workflow-definition type family in @bakin/core
Extract WorkflowDefinition and its step/input/layout siblings from
plugins/workflows/types.ts into packages/core/src/workflows/
definition-types.ts so core consumers (plugin registry, agent-package
loader, package-integrity) stop importing the workflows plugin's source
tree for types. The plugin re-exports them type-only (erased at build;
same pattern as the tasks-plugin board types). WorkflowDefinition is now
pinned to its SDK + core homes by the type-single-home scan.
