+++ packages/core/package.json
@@ -49,6 +49,7 @@
     "./storage/db": "./src/storage/db.ts",
     "./execution/ledger": "./src/execution/ledger.ts",
     "./events": "./src/events/event-bus.ts",
+    "./workflows/*": "./src/workflows/*.ts",
     "./parsers/*": "./src/parsers/*.ts",
     "./format": "./src/format.ts"
   },
