+++ src/core/agent-packages/package-integrity.ts
@@ -3,7 +3,7 @@ import { splitFrontmatter } from '@bakin/core/format/frontmatter'
 import { basename, isAbsolute, normalize, relative, resolve } from 'path'
 import yaml from 'js-yaml'
 import type { Manifest } from '../../../packages/core/src/agent-packages/manifest'
-import type { WorkflowDefinition } from '../../../plugins/workflows/types'
+import type { WorkflowDefinition } from '@bakin/core/workflows/definition-types'
 import { validateDefinition } from '../../../plugins/workflows/lib/parser'
 import { validatePackageLessonIntegrity } from './lesson-integrity'
